% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_data_cleanup.R
\name{draw_graph}
\alias{draw_graph}
\title{Draw a histogram...}
\usage{
draw_graph(x)
}
\arguments{
\item{x}{The list output from \link{likert_scale_analyzer}.}
}
\description{
Draw a histogram...
}
\examples{
# likert_results <- likert_scale_analyzer(
#   data = religious_optimism,
#   likert_cols = c("relig_practice0", "relig_q4","relig_q5"),
#   invalid_values = c(" ", "NA")
#   )
#  draw_graph(likert_results)

}
