% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_data_cleanup.R
\name{print.Likert_List}
\alias{print.Likert_List}
\title{Define print method for results_list using S3 method: print.likert_scale()}
\usage{
\method{print}{Likert_List}(x)
}
\arguments{
\item{x}{Any list object belonging to the "Likert_list" class}
}
\description{
Define print method for results_list using S3 method: print.likert_scale()
}
