% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_data_cleanup.R
\name{validate_likert}
\alias{validate_likert}
\title{MISSING TITLE}
\usage{
validate_likert(x)
}
\arguments{
\item{x}{the argument received by the \code{draw_graph()} function}
}
\value{
x
}
\description{
MISSING TITLE
}
\examples{
likert_results <- likert_scale_analyzer(
  data = religious_som,
  likert_cols = c("relig_practice0", "relig_q4","relig_q5"),
  invalid_values = c(" ", "NA"))

 draw_graph(likert_results)

}
