% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_data_cleanup.R
\name{as.data.frame.Likert_List}
\alias{as.data.frame.Likert_List}
\title{Define as.data.frame method for results_list using S3 method: as.data.frame.likert_scale()}
\usage{
\method{as.data.frame}{Likert_List}(list_results)
}
\arguments{
\item{list_results}{Any list object belonging to the "Likert_List" class}
}
\description{
Define as.data.frame method for results_list using S3 method: as.data.frame.likert_scale()
}
