% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_data_cleanup.R
\name{draw_graph}
\alias{draw_graph}
\title{Draw a histogram}
\usage{
draw_graph(x)
}
\arguments{
\item{x}{The dataframe created when \link{likert_scale_analyzer} is called.}
}
\value{
A series of facet_wrap histogram, corresponding to the likert_scale.
}
\description{
It plots a series of histogram summarizing valid response counts by each question, differentiating by likert scale.
}
\examples{
# likert_results <- likert_scale_analyzer(
#   data = religious_som,
#   likert_cols = c("relig_practice0", "relig_q4","relig_q5"),
#   invalid_values = c(" ", "NA")
#   )
#  draw_graph(likert_results)

}
