% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_data_cleanup.R
\name{likert_scale_analyzer}
\alias{likert_scale_analyzer}
\title{Use likert_scale_analyzer()}
\usage{
likert_scale_analyzer(data, likert_cols, invalid_values = c(" ", "NA", "N/A"))
}
\arguments{
\item{data}{A data frame containing survey questions and responses uploaded by the user.}

\item{likert_cols}{A character vector containing Likert-scale question names in c("","", ...).}

\item{invalid_values}{A character vector containing invalid values such as empty string, NA values, N/A
values, which are pre-set.}
}
\value{
A list object belonging to the \code{Likert_List} class where each element contains a nested list with
the following 6 components:
\itemize{
\item variable_name$question: A character vector
\item variable_name$valid_count: A scalar numeric vector
\item variable_name$invalid_count: A scalar numeric vector, may be empty
\item variable_name$scale_min: A scalar numeric vector
\item variable_name$scale_max: A scalar numeric vector
\item variable_name$response_counts: A table that includes scale point and valid counts
}
}
\description{
Given a likert scale survey, the \code{likert_scale_analyzer} function filters out invalid responses
such as NA values and empty strings, calculates scale points and range for each question, and
returns a list object holding counts for valid and invalid responses.
}
\examples{
likert_scale_analyzer(
  data = religious_som,
  likert_cols = c("relig_q4","relig_experience1","relig_experience4","SOM_q1","SOM_q3"),
  invalid_values = c(" ", "NA", "N/A")
 )
}
