% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_data_cleanup.R
\name{likert_scale_analyzer}
\alias{likert_scale_analyzer}
\title{Use the xkcd API}
\usage{
likert_scale_analyzer(data, likert_cols, invalid_values = c(" ", "NA", "N/A"))
}
\description{
Given a comic number, the \code{xkcd()} function calls the xkcd JSON API and returns metadata about the comic in the form of a list object.
}
