% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_data_cleanup.R
\name{validate_likert}
\alias{validate_likert}
\title{Validating objects to ensure they belong to the \code{Likert_List} class}
\usage{
validate_likert(x)
}
\arguments{
\item{x}{An object that is going to be validated to ensure that it belongs to the \code{Likert_List} class.}
}
\value{
If object \code{x} does belong to the \code{Likert_List} class, \code{validate_likert()} will return the object.
Otherwise, it will throw an error.
}
\description{
The validator function \code{validate_likert()} certifies that the argument \code{x} belongs to the
\code{Likert_List} class. It is used within the \code{\link[=draw_graph]{draw_graph()}} function to validate input.
}
\examples{
likert_results <- likert_scale_analyzer(
  data = religious_som,
  likert_cols = c("relig_practice0", "relig_q4","relig_q5"),
  invalid_values = c(" ", "NA"))

 draw_graph(likert_results)
}
